nasasite/requirements.txt
django
django-filer
django-admin-sortable2
python-dotenv
mysqlclient